Update Native S3 Samples to use explicit credentials + dataset
As we're reworking Hyper's internals of the experimental s3 feature,
we now require the user to explicitly specify credentials in a
s3-location constructor - even for public resources that don't require
authentication.

While working on this, we discovered that the nyc tlc taxi data set
we used in one example is currently inaccessible, therefore we change
this example to use one of our internal parquet datasets.