Remove version pinning from git-to-hyper example
Pinned version might lead to the case that users are using outdated
versions with known vulnerabilities, as we do not guarantee active
maintenance of Community Supported examples. As Salesforce values Trust
and Security above all other things we rather accept that our example
might break with future releases of our dependencies than risking that
our users are making themselves vulnerable because of us.